+++ packages/jsii-pacmak/test/generated-code/harness.ts
@@ -1,3 +1,4 @@
+import { createPatch } from 'diff';
 import * as fs from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
@@ -6,37 +7,51 @@ import * as process from 'process';
 import { pacmak, TargetName } from '../../lib';
 import { shell } from '../../lib/util';
 
+export const JSII_TEST_PACKAGES: readonly string[] = [
+  '@scope/jsii-calc-base-of-base',
+  '@scope/jsii-calc-base',
+  '@scope/jsii-calc-lib',
+  'jsii-calc',
+];
+
+const DIFF = Symbol('diff');
 const FILE = Symbol('file');
 const MISSING = Symbol('missing');
 const TARBALL = Symbol('tarball');
 const BINARY = Symbol('binary');
 export const TREE = Symbol('tree');
+const TREE_ROOT = Symbol('treeRoot');
 
 // Custom serializers so we can see the source without escape sequences
 expect.addSnapshotSerializer({
-  test: (val) => val?.[FILE] != null,
+  test: (val) => DIFF in val,
+  serialize: (val) => val[DIFF],
+});
+
+expect.addSnapshotSerializer({
+  test: (val) => FILE in val,
   serialize: (val) => val[FILE],
 });
 expect.addSnapshotSerializer({
-  test: (val) => val?.[MISSING] != null,
+  test: (val) => MISSING in val,
   serialize: (val) => `${val[MISSING]} does not exist`,
 });
 
 expect.addSnapshotSerializer({
-  test: (val) => val?.[TARBALL] != null,
+  test: (val) => TARBALL in val,
   serialize: (val) =>
     `${val[TARBALL]} ${
       val[TARBALL].endsWith('.tgz') ? 'is' : 'embeds'
     } a tarball`,
 });
 expect.addSnapshotSerializer({
-  test: (val) => val?.[BINARY] != null,
+  test: (val) => BINARY in val,
   serialize: (val) => `${val[BINARY]} is a binary file`,
 });
 expect.addSnapshotSerializer({
-  test: (val) => val?.[TREE] != null,
+  test: (val) => TREE in val,
   serialize: (val) => {
-    return `<root>\n${formatTree(val[TREE])}`;
+    return `${val[TREE_ROOT] ?? '<root>'}\n${formatTree(val[TREE])}`;
   },
 });
 
@@ -57,25 +72,40 @@ export function verifyGeneratedCodeFor(
     done();
   });
 
-  for (const pkg of [
-    '@scope/jsii-calc-base-of-base',
-    '@scope/jsii-calc-base',
-    '@scope/jsii-calc-lib',
-    'jsii-calc',
-  ]) {
+  for (const pkg of JSII_TEST_PACKAGES) {
     // Extend timeout, because this could be slow (python has more time because of the mypy pass)...
     jest.setTimeout(timeout);
 
     test(`Generated code for ${JSON.stringify(pkg)}`, async () => {
       const pkgRoot = path.resolve(__dirname, '..', '..', '..', pkg);
-      await runPacmak(pkgRoot, targetName, outDir);
 
-      expect({ [TREE]: checkTree(outDir) }).toMatchSnapshot('<outDir>/');
-
-      if (targetName !== TargetName.PYTHON || process.env.SKIP_MYPY_CHECK) {
-        return Promise.resolve();
+      const outDirRuntimeTypeChecked = path.join(
+        outDir,
+        'runtime-type-checking',
+      );
+      await runPacmak(pkgRoot, targetName, outDirRuntimeTypeChecked);
+      expect({ [TREE]: checkTree(outDirRuntimeTypeChecked) }).toMatchSnapshot(
+        '<outDir>/',
+      );
+      if (targetName === TargetName.PYTHON && !process.env.SKIP_MYPY_CHECK) {
+        await runMypy(path.join(outDirRuntimeTypeChecked, targetName));
       }
-      return runMypy(path.join(outDir, targetName));
+
+      // Now we'll generate WITHOUT runtime type-checks, and assert on the differences
+      const outDirNotRuntimeTypeChecked = path.join(
+        outDir,
+        'no-runtime-type-checking',
+      );
+      await runPacmak(pkgRoot, targetName, outDirNotRuntimeTypeChecked, {
+        runtimeTypeChecking: false,
+      });
+      expect({
+        [TREE]: diffTrees(
+          outDirRuntimeTypeChecked,
+          outDirNotRuntimeTypeChecked,
+        ),
+        [TREE_ROOT]: '<runtime-type-check-diff>',
+      }).toMatchSnapshot('<runtime-type-check-diff>/');
     });
   }
 }
@@ -129,16 +159,82 @@ export function checkTree(
       }
       return tree;
     }, {} as { [name: string]: TreeStructure });
+}
 
-  function tryStat(at: string) {
-    try {
-      return fs.statSync(at);
-    } catch (e: any) {
-      if (e.code !== os.constants.errno.ENOENT) {
-        throw e;
-      }
+export function diffTrees(
+  original: string,
+  updated: string,
+  { root = original }: { root?: string } = {},
+): TreeStructure | undefined {
+  const originalStat = tryStat(original);
+  const updatedStat = tryStat(updated);
+
+  // Should exist on both sides AND have the same file type
+  expect(originalStat?.isDirectory()).toBe(updatedStat?.isDirectory());
+  expect(originalStat?.isFile()).toBe(updatedStat?.isFile());
+
+  const relativeFile = path.relative(root, original).replace(/\\/g, '/');
+  if (originalStat?.isFile()) {
+    if (original.endsWith('.tgz') || original.endsWith('.png')) {
+      // This is a binary object, these should match exactly
+      expect(fs.readFileSync(original)).toEqual(fs.readFileSync(updated));
       return undefined;
     }
+    const patch = createPatch(
+      relativeFile,
+      fs.readFileSync(original, 'utf-8'),
+      fs.readFileSync(updated, 'utf-8'),
+      '--runtime-type-checking',
+      '--no-runtime-type-checking',
+      {
+        context: 5,
+        ignoreWhitespace: false,
+        newlineIsToken: true,
+      },
+    )
+      .trimEnd()
+      .split(/\n/)
+      .slice(2);
+
+    if (patch.length === 2) {
+      return undefined;
+    }
+
+    const snapshotName = `<runtime-type-check-diff>/${relativeFile}.diff`;
+    expect({ [DIFF]: patch.join('\n') }).toMatchSnapshot(snapshotName);
+    return `${path.basename(original)}.diff`;
+  }
+
+  return fs
+    .readdirSync(original)
+    .map((entry) => ({
+      entry,
+      subtree: diffTrees(
+        path.join(original, entry),
+        path.join(updated, entry),
+        { root },
+      ),
+    }))
+    .reduce((tree, { entry, subtree }) => {
+      if (subtree != null) {
+        tree = tree ?? {};
+        if (typeof subtree === 'string' && subtree.startsWith(entry)) {
+          entry = subtree;
+        }
+        tree[entry] = subtree;
+      }
+      return tree;
+    }, undefined as { [name: string]: TreeStructure } | undefined);
+}
+
+function tryStat(at: string) {
+  try {
+    return fs.statSync(at);
+  } catch (e: any) {
+    if (e.code !== os.constants.errno.ENOENT) {
+      throw e;
+    }
+    return undefined;
   }
 }
 
@@ -146,6 +242,9 @@ async function runPacmak(
   root: string,
   targetName: TargetName,
   outdir: string,
+  {
+    runtimeTypeChecking = true,
+  }: { readonly runtimeTypeChecking?: boolean } = {},
 ): Promise<void> {
   return expect(
     pacmak({
@@ -153,13 +252,22 @@ async function runPacmak(
       fingerprint: false,
       inputDirectories: [root],
       outputDirectory: outdir,
+      runtimeTypeChecking,
       targets: [targetName],
     }),
   ).resolves.not.toThrowError();
 }
 
-async function runMypy(pythonRoot: string): Promise<void> {
-  const venvRoot = path.join(__dirname, '.venv');
+export async function preparePythonVirtualEnv({
+  install = [],
+  venvDir = __dirname,
+  systemSitePackages = true,
+}: {
+  install?: readonly string[];
+  venvDir?: string;
+  systemSitePackages?: boolean;
+} = {}) {
+  const venvRoot = path.join(venvDir, '.venv');
   const venvBin = path.join(
     venvRoot,
     process.platform === 'win32' ? 'Scripts' : 'bin',
@@ -184,11 +292,16 @@ async function runMypy(pythonRoot: string): Promise<void> {
     shell(process.platform === 'win32' ? 'python' : 'python3', [
       '-m',
       'venv',
-      '--system-site-packages', // Allow using globally installed packages (saves time & disk space)
+      ...(systemSitePackages
+        ? [
+            '--system-site-packages', // Allow using globally installed packages (saves time & disk space)
+          ]
+        : []),
       JSON.stringify(venvRoot),
     ]),
   ).resolves.not.toThrowError();
-  // Install mypy and the jsii runtime in there as needed
+
+  // Install development dependencies as needed...
   await expect(
     shell(
       venvPython,
@@ -199,6 +312,8 @@ async function runMypy(pythonRoot: string): Promise<void> {
         '--no-input',
         '-r',
         path.resolve(__dirname, 'requirements-dev.txt'),
+        // Additional install parameters
+        ...install,
         // Note: this resolution is a little ugly, but it's there to avoid creating a dependency cycle
         JSON.stringify(
           path.resolve(
@@ -210,6 +325,13 @@ async function runMypy(pythonRoot: string): Promise<void> {
       { env, retry: { maxAttempts: 5 } },
     ),
   ).resolves.not.toThrowError();
+
+  return { env, venvPython, venvRoot };
+}
+
+async function runMypy(pythonRoot: string): Promise<void> {
+  const { env, venvPython } = await preparePythonVirtualEnv();
+
   // Now run mypy on the Python code
   return expect(
     shell(
@@ -244,7 +366,11 @@ async function runMypy(pythonRoot: string): Promise<void> {
 
 type TreeStructure = string | { [name: string]: TreeStructure };
 
-function formatTree(tree: TreeStructure): string {
+function formatTree(tree: TreeStructure | undefined): string {
+  if (tree == null) {
+    return `┗━ 🕳 There is nothing here`;
+  }
+
   if (typeof tree === 'string') {
     return `┗━ 📄 ${tree}`;
   }